+++ lib/userstamp/model_concern.rb
@@ -1,6 +1,0 @@
-module Userstamp::ModelConcern
-  extend ActiveSupport::Concern
-
-  include Userstamp::Stampable
-  include Userstamp::Stamper
-end