+++ lib/userstamper/controller_concern.rb
@@ -1,10 +1,10 @@
-# The +Userstamp::Controller+ module, when included into a controller, adds an +before_action+
+# The +Userstamper::Controller+ module, when included into a controller, adds an +before_action+
 # callback (named +set_stamper+) and an +after_action+ callback (named +reset_stamper+). These
 # methods assume a couple of things, but can be re-implemented in your controller to better suit
 # your application.
 #
 # See the documentation for `set_stamper` and `reset_stamper` for specific implementation details.
-module Userstamp::ControllerConcern
+module Userstamper::ControllerConcern
   extend ActiveSupport::Concern
 
   included do
@@ -30,15 +30,15 @@ module Userstamp::ControllerConcern
   # your application you will want to manually add your own implementation of this method to the
   # private section of your +ApplicationController+
   def set_stamper
-    @_userstamp_stamper = Userstamp.config.default_stamper_class.stamper
-    Userstamp.config.default_stamper_class.stamper = current_user
+    @_userstamp_stamper = Userstamper.config.default_stamper_class.stamper
+    Userstamper.config.default_stamper_class.stamper = current_user
   end
 
   # The {#reset_stamper} method as implemented here assumes that a +User+ model is being used as
   # the stamper. If this is not the case then you will need to manually add your own
   # implementation of this method to the private section of your +ApplicationController+
   def reset_stamper
-    Userstamp.config.default_stamper_class.stamper = @_userstamp_stamper
+    Userstamper.config.default_stamper_class.stamper = @_userstamp_stamper
     @_userstamp_stamper = nil
   end
 end