+++ lib/userstamper/model_concern.rb
@@ -1,0 +1,6 @@
+module Userstamper::ModelConcern
+  extend ActiveSupport::Concern
+
+  include Userstamper::Stampable
+  include Userstamper::Stamper
+end