+++ lib/userstamper/utilities.rb
@@ -1,4 +1,4 @@
-module Userstamp::Utilities
+module Userstamper::Utilities
   # Removes the association methods from the model.
   #
   # @param [Class] model The model to remove methods from.
@@ -20,20 +20,18 @@ module Userstamp::Utilities
     end
   end
 
-  # Obtains the creator/updater/deleter columns which are present in the model.
+  # Obtains the creator/updater columns which are present in the model.
   #
   # @param [Class] model The model to query.
   # @return [nil|Array<(bool, bool, bool)>] Nil if the model does not have a table defined.
-  #   Otherwise, a tuple of booleans indicating the presence of the created, updated, and deleted
-  #   columns.
+  #   Otherwise, a tuple of booleans indicating the presence of the created, updated columns.
   def self.available_association_columns(model)
     return nil if model.name.nil? || model.table_name.empty?
     columns = Set[*model.column_names]
-    config = Userstamp.config
+    config = Userstamper.config
 
     [config.creator_attribute.present? && columns.include?(config.creator_attribute.to_s),
-     config.updater_attribute.present? && columns.include?(config.updater_attribute.to_s),
-     config.deleter_attribute.present? && columns.include?(config.deleter_attribute.to_s)]
+     config.updater_attribute.present? && columns.include?(config.updater_attribute.to_s)]
   rescue ActiveRecord::StatementInvalid => _
     nil
   end