+++ spec/lib/configuration_spec.rb
@@ -4,16 +4,16 @@ RSpec.describe 'Configuration' do
   describe '.configure' do
     it 'yields a block with the config argument' do
       block = nil
-     Userstamp.configure do |config|
+     Userstamper.configure do |config|
         block = config
       end
-      expect(block).to be(Userstamp::Configuration)
+      expect(block).to be(Userstamper::Configuration)
     end
   end
 
   describe '.config' do
     it 'matches default_stamper and default_stamper_class' do
-      config = Userstamp.config
+      config = Userstamper.config
       expect(config.default_stamper_class.name).to eq(config.default_stamper)
     end
   end