+++ CHANGELOG.md
@@ -1,6 +1,6 @@
  # Changelog
  ## 4.0.0 (09-04-2021)
-  * Priit Tark        - Rails 6.1 support
+  * Priit Tark        - Rails 6.1 support and renamed gem to userstamper
 
  ## 3.0.5 (22-8-2017)
   * Jonathan Putney   - Support Rails 5.0.